+++ src/App.tsx
@@ -1,10 +1,12 @@
 import './App.scss'
 
-import Collapse from '../' // From dist folder
+import { Collapse } from '../lib/Collapse' // From library source
 import React from 'react'
 import cx from 'classnames'
 
-//import Collapse from '../lib/Collapse' // From library source
+// import { Collapse } from '../' // From dist folder
+// import { Collapse } from '../dist/Collapse.es' // From dist folder
+//import { Collapse } from '../lib/Collapse' // From library source
 
 class App extends React.Component {
   state = {